refactor: change folder name patterns
