+++ .gitignore
@@ -10,7 +10,6 @@ pnpm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 lerna-debug.log*
-!/src/logs
 
 # OS
 .DS_Store
