+++ src/app.module.ts
@@ -1,22 +1,22 @@
 import { Module } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
 
-import { LogConfig } from './logs/log.config';
-import { LogsController } from './logs/controllers/log.controller';
-import { LogsService } from './logs/services/log.service';
-import { LogRepository } from './logs/log.repository';
-import { LogSchema } from './logs/schemas/log.schema';
+import { LogConfig } from './log/log.config';
+import { LogsController } from './log/controllers/log.controller';
+import { LogsService } from './log/services/log.service';
+import { LogRepository } from './log/log.repository';
+import { LogSchema } from './log/schemas/log.schema';
 import { DatabaseConfig } from './common/database/database.config';
-import { MachineConfig } from './machines/machine.config';
-import { MachineSchema } from './machines/schemas/machine.schema';
-import { MachinesController } from './machines/controllers/machine.controller';
-import { MachinesService } from './machines/services/machine.service';
-import { MachineRepository } from './machines/machine.repository';
-import { EstablishmentConfig } from './establishments/establishment.config';
-import { EstablishmentSchema } from './establishments/schemas/establishment.schema';
-import { EstablishmentsController } from './establishments/controllers/establishment.controller';
-import { EstablishmentsService } from './establishments/services/establishment.service';
-import { EstablishmentRepository } from './establishments/establishment.repository';
+import { MachineConfig } from './machine/machine.config';
+import { MachineSchema } from './machine/schemas/machine.schema';
+import { MachinesController } from './machine/controllers/machine.controller';
+import { MachinesService } from './machine/services/machine.service';
+import { MachineRepository } from './machine/machine.repository';
+import { EstablishmentConfig } from './establishment/establishment.config';
+import { EstablishmentSchema } from './establishment/schemas/establishment.schema';
+import { EstablishmentsController } from './establishment/controllers/establishment.controller';
+import { EstablishmentsService } from './establishment/services/establishment.service';
+import { EstablishmentRepository } from './establishment/establishment.repository';
 
 @Module({
   imports: [
