+++ src/log/dto/log.dto.ts
@@ -1,6 +1,6 @@
 import * as validator from 'class-validator';
-import { Machine } from '../../machines/interfaces/machine.interface';
-import { Establishment } from '../../establishments/interfaces/establishment.interface';
+import { Machine } from '../../machine/interfaces/machine.interface';
+import { Establishment } from '../../establishment/interfaces/establishment.interface';
 
 export class LogDTO {
   @validator.IsNotEmpty()
