+++ src/log/interfaces/log.interface.ts
@@ -1,6 +1,6 @@
 import { Document } from 'mongoose';
-import { Machine } from '../../machines/interfaces/machine.interface';
-import { Establishment } from '../../establishments/interfaces/establishment.interface';
+import { Machine } from '../../machine/interfaces/machine.interface';
+import { Establishment } from '../../establishment/interfaces/establishment.interface';
 
 export interface Log extends Document {
   readonly machineId: Machine;
