+++ src/log/schemas/log.schema.ts
@@ -1,8 +1,8 @@
 import { Schema, Prop, SchemaFactory } from '@nestjs/mongoose';
 
 import * as mongoose from 'mongoose';
-import { Machine } from '../../machines/interfaces/machine.interface';
-import { Establishment } from '../../establishments/interfaces/establishment.interface';
+import { Machine } from '../../machine/interfaces/machine.interface';
+import { Establishment } from '../../establishment/interfaces/establishment.interface';
 
 @Schema({ versionKey: false })
 class Log {
