+++ src/machine/machine.config.ts
