+++ src/machine/machine.repository.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 
 import { Model } from 'mongoose';
-import { Establishment } from '../establishments/interfaces/establishment.interface';
+import { Establishment } from '../establishment/interfaces/establishment.interface';
 import { MachineDTO } from './dto/machine.dto';
 import { Machine } from './interfaces/machine.interface';
 import { MachineConfig } from './machine.config';
