+++ src/machine/schemas/machine.schema.ts
@@ -1,7 +1,7 @@
 import { Schema, Prop, SchemaFactory } from '@nestjs/mongoose';
 
 import * as mongoose from 'mongoose';
-import { Establishment } from '../../establishments/interfaces/establishment.interface';
+import { Establishment } from '../../establishment/interfaces/establishment.interface';
 
 @Schema({ versionKey: false })
 class Machine {
